Apply the Choi algorithm

The Choi algorithm detects periods of non-wear in activity data from an ActiGraph device. Such intervals are likely to represent invalid data and therefore should be excluded from downstream analysis.

apply_choi(
  agdb,
  min_period_len = 90,
  min_window_len = 30,
  spike_tolerance = 2,
  use_magnitude = FALSE
)

Arguments

agdb
A tibble (tbl) of activity data (at least) an epochlength attribute. The epoch length must be 60 seconds.

min_period_len
Minimum number of consecutive "zero" epochs to start a non-wear period. The default is 90.

min_window_len
The minimum number of consecutive "zero" epochs immediately preceding and following a spike of artifactual movement. The default is 30.

spike_tolerance
Also known as artifactual movement interval. At most spike_tolerance "nonzero" epochs can occur in sequence during a non-wear period without interrupting it. The default is 2.

use_magnitude
Logical. If true, the magnitude of the vector (axis1, axis2, axis3) is used to measure activity; otherwise the axis1 value is used. The default is FALSE.

Value

A summary tibble of the detected non-wear periods. If the activity data is grouped, then non-wear periods are detected separately for each group.

Details

The Choi algorithm extends the Troiano algorithm by requiring that short spikes of artifactual movement during a non-wear period are preceded and followed by min_window_len consecutive "zero" epochs.

This implementation of the algorithm expects that the epochs are 60 second long.

References

L Choi, Z Liu, CE Matthews and MS Buchowski. Validation of accelerometer wear and nonwear time classification algorithm. Medicine & Science in Sports & Exercise, 43(2):357–364, 2011.

ActiLife 6 User's Manual by the ActiGraph Software Department. 04/03/2012.

See also

Examples

library("dplyr")
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
data("gtxplus1day")

gtxplus1day %>%
  collapse_epochs(60) %>%
  apply_choi()
#>   period_start          period_end length
#> 1   2012-06-28 2012-06-28 02:37:00    157