Apply the Cole-Kripke algorithm

The Cole-Kripke sleep scoring algorithm is primarily used for adult populations as the supporting research was performed on subjects ranging from 35 to 65 years of age.

apply_cole_kripke(agdb)

Arguments

agdb
A tibble (tbl) of activity data (at least) an epochlength attribute. The epoch length must be 60 seconds.

Value

A tibble (tbl) of activity data. A new column sleep indicates whether each 60s epoch is scored as asleep (S) or awake (W).

Details

The original paper proposes three versions of the Cole-Kripke algorithm, optimized for 1-minute, 30-second and 10-second epochs. Here only the 1-min version is implemented and therefore the apply_cole_kripke function requires that the activity data is in 60s epochs. Use the collapse_epochs function to modify higher-frequency data, if necessary.

The Cole-Kripke algorithm uses the y-axis (axis 1) counts. First epoch counts are divided by 100 and any scaled counts over 300 are clipped to 300. This transformation is specific to ActiGraph devices. The sleep index (SI) is defined as

.001 * (106 * epoch_prev(4) + 54 * epoch_prev(3) + 58 * epoch_prev(2) + 76 * epoch_prev(1) + 230 * epoch + 74 * epoch_next(1) + 67 * epoch_next(2))

where at epoch t, epoch_prev(i) is the scaled activity count i epochs before t. Similarly, epoch_next(i) is the scaled activity count i epochs before t. That is, the algorithm uses a 7-epoch window which includes the four preceding and the two subsequent epochs. The time series of activity counts is padded with zeros as necessary, at the beginning and at the end.

Finally, the sleep state is awake (W) if the sleep index SI is less than 1; otherwise the sleep state is asleep (S).

References

RJ Cole, DF Kripke, W Gruen, DJ Mullaney and JC Gillin. Automatic sleep/wake identification from wrist activity. Sleep, 15(5):461–469, 1992.

ActiLife 6 User's Manual by the ActiGraph Software Department. 04/03/2012.

See also

Examples

library("dplyr")
data("gtxplus1day")

gtxplus1day %>%
  collapse_epochs(60) %>%
  apply_cole_kripke()
#> # A tibble: 1,500 x 6
#>    timestamp           axis1 axis2 axis3 count sleep
#>    <dttm>              <int> <int> <int> <dbl> <chr>
#>  1 2012-06-27 10:54:00  1465  1791  2572 14.6  W    
#>  2 2012-06-27 10:55:00   207   218   270  2.07 W    
#>  3 2012-06-27 10:56:00   169   257   270  1.69 W    
#>  4 2012-06-27 10:57:00     0     0     0  0    W    
#>  5 2012-06-27 10:58:00   157   174   248  1.57 W    
#>  6 2012-06-27 10:59:00    23    23   279  0.23 S    
#>  7 2012-06-27 11:00:00     0     0     0  0    S    
#>  8 2012-06-27 11:01:00     0     0     0  0    S    
#>  9 2012-06-27 11:02:00     0     0     0  0    S    
#> 10 2012-06-27 11:03:00     0     0     0  0    S    
#> # … with 1,490 more rows