Apply the Tudor-Locke algorithm

The Tudor-Locke algorithm detects periods of time in bed and, for each period, computes sleep quality metrics such as total minutes in bed, total sleep time, number and average length of awakenings, movement and fragmentation index.

apply_tudor_locke(
  agdb,
  n_bedtime_start = 5,
  n_wake_time_end = 10,
  min_sleep_period = 160,
  max_sleep_period = 1440,
  min_nonzero_epochs = 0
)

Arguments

agdb
A tibble (tbl) of activity data (at least) an epochlength attribute. The epoch length must be 60 seconds. Each epoch should be scored as asleep (S) or awake (W), using the Sadeh, the Cole-Kripke or a custom algorithm.

n_bedtime_start
Bedtime definition, in minutes. The default is 5.

n_wake_time_end
Wake time definition, in minutes. The default is 10.

min_sleep_period
Minimum sleep period length, in minutes. The default is 160.

max_sleep_period
Maximum sleep period length, in minutes. The default is 1440 (24 hours).

min_nonzero_epochs
Minimum number of epochs with non-zero activity. The default is 0.

Value

A summary tibble of the detected sleep periods. If the activity data is grouped, then sleep periods are detected separately for each group.

in_bed_time

The first minute of the bedtime.

out_bed_time

The first minute of wake time.

onset

The first minute that the algorithm scores "asleep".

latency

The time elapsed between bedtime and sleep onset. By the definition of Tudor-Locke, latency is 0.

efficiency

The number of sleep minutes divided by the bedtime minutes.

duration

The duration of the sleep period, in minutes.

activity_counts

The sum of activity counts for the entire sleep period.

total_sleep_time

The number of minutes scored as "asleep" during the sleep period.

wake_after_onset

The number of minutes scored as "awake", minus the sleep latency, during the sleep period.

nb_awakenings

The number of awakening episodes.

ave_awakening

The average length, in minutes, of all awakening episodes.

movement_index

Proportion of awake time out of the total time in bed, in percentages.

fragmentation_index

Proportion of one-minute sleep bouts out of the number of sleep bouts of any length, in percentages.

sleep_fragmentation_index

The sum of the movement and fragmentation indices.

nonzero_epochs

The number of epochs with activity > 0 (nonzero epochs).

Details

Once each one-minute epoch is labeled as asleep (S) or awake (W), we can use the Tudor-Locke algorithm to detect periods of bedtime and sleep time. By definition, sleep time < bedtime since one can be in bed and not sleeping.

Bedtime is (the first minute of) n_bedtime_start consecutive epochs/minutes labeled asleep (S). Similarly, wake time is (the first minute of) of n_wake_time_end consecutive epochs/minutes labeled awake (W), after a period of sleep. The block of time between bedtime and wake time is one sleep period, if the time elapsed is at least min_sleep_period minutes. There can be multiple sleep periods in 24 hours but a sleep period cannot be longer than max_sleep_period minutes.

For each sleep period, the algorithm calculates several measures of sleep quality such as time asleep and time awake, number and average length of awakenings, and movement and fragmentation indices.

This implementation of the Tudor-Locke algorithm detects all the sleep periods that ActiLife detects and, in some cases, it detects additional sleep periods. There are (at least) two such cases:

ActiLife filters out some sleep periods with exactly min_nonzero_epochs nonzero epochs. Alternatively, ActiLife computes the number of nonzero epochs in a sleep period differently and sometimes underestimates nonzero_epochs compared to apply_tudor_locke.

ActiLife filters out sleep periods that end when the activity data ends, i.e., when the out_bed_time is also the final timestamp in the agdb table.

References

C Tudor-Locke, TV Barreira, JM Schuna Jr, EF Mire and PT Katzmarzyk. Fully automated waist-worn accelerometer algorithm for detecting children's sleep-period time separate from 24-h physical activity or sedentary behaviors. Applied Physiology, Nutrition, and Metabolism, 39(1):53–57, 2014.

ActiLife 6 User's Manual by the ActiGraph Software Department. 04/03/2012.

See also

Examples

library("dplyr")
library("lubridate")
#> 
#> Attaching package: ‘lubridate’
#> The following objects are masked from ‘package:base’:
#> 
#>     date, intersect, setdiff, union
data("gtxplus1day")

# Detect sleep periods using Sadeh as the sleep/awake algorithm
# and Tudor-Locke as the sleep period algorithm
agdb <- gtxplus1day %>%
  collapse_epochs(60) %>%
  filter(day(timestamp) == 28)
periods_sleep <- agdb %>%
  apply_sadeh() %>%
  apply_tudor_locke(min_sleep_period = 60)
periods_sleep
#>           in_bed_time        out_bed_time               onset latency
#> 1 2012-06-28 00:00:00 2012-06-28 07:38:00 2012-06-28 00:00:00       0
#> 2 2012-06-28 09:41:00 2012-06-28 10:53:00 2012-06-28 09:41:00       0
#>   efficiency duration activity_counts nonzero_epochs total_sleep_time
#> 1   97.16157      458            9126             27              445
#> 2   84.72222       72           10662             24               61
#>   wake_after_onset nb_awakenings ave_awakening movement_index
#> 1               13             4          3.25       5.895197
#> 2               11             2          5.50      33.333333
#>   fragmentation_index sleep_fragmentation_index
#> 1                  40                  45.89520
#> 2                   0                  33.33333

# Group and summarize by an extra varible (hour < 6 or not), which
# splits one long sleep period in two
agdb <- agdb %>%
  mutate(hour_lt6 = hour(timestamp) < 6) %>%
  group_by(hour_lt6)
periods_sleep <- agdb %>%
  apply_sadeh() %>%
  apply_tudor_locke(min_sleep_period = 60)
periods_sleep
#> # A tibble: 3 x 16
#> # Groups:   hour_lt6 [2]
#>   hour_lt6 in_bed_time         out_bed_time        onset               latency
#>   <lgl>    <dttm>              <dttm>              <dttm>                <int>
#> 1 FALSE    2012-06-28 06:00:00 2012-06-28 07:38:00 2012-06-28 06:00:00       0
#> 2 FALSE    2012-06-28 09:41:00 2012-06-28 10:53:00 2012-06-28 09:41:00       0
#> 3 TRUE     2012-06-28 00:00:00 2012-06-28 06:00:00 2012-06-28 00:00:00       0
#> # … with 11 more variables: efficiency <dbl>, duration <int>,
#> #   activity_counts <int>, nonzero_epochs <int>, total_sleep_time <int>,
#> #   wake_after_onset <int>, nb_awakenings <int>, ave_awakening <dbl>,
#> #   movement_index <dbl>, fragmentation_index <dbl>,
#> #   sleep_fragmentation_index <dbl>