Re-integrate epochs

Collapse post-filtered activity counts into larger epoch "buckets".

collapse_epochs(agdb, epoch_len_out, use_incomplete = TRUE)

Arguments

agdb
A tibble (tbl) of activity data (at least) an epochlength attribute.

epoch_len_out
Output (longer) epoch length in seconds, must be exact multiple of the input epoch length. Currently only epoch_len_out = 60 is supported.

use_incomplete
logical. Set to TRUE to follow ActiLife convention, which collapses all observed epochs even if they are incomplete.

Value

A tibble (tbl) of activity data collapsed into one-minute epochs.

Details

Activity counts cannot be reintegrated into shorter epochs, e.g., 60s -> 10s. Currently, collapse_epochs integrates into 60s epochs only. This is not general but is sufficient for sleep analysis because the standard Sadeh and Cole-Kripke sleep algorithms were developed for 60s epoch data.

Suppose we want to collapse from 15 to 60 seconds. A complete 60s epoch consists of four 15s epochs: 00, 15, 45 and 60. However, the first and last epochs would be incomplete if the device started/stopped collecting data mid-minute. ActiLife 6 uses these epochs anyway. For example, if only 45 and 60 are available for the first minute, then ActiLife will aggregate across these two epochs only. This is a reasonable approach to sleep analysis with the Sadeh and the Cole-Kripke algorithms which pad the beginning and the end of the time series with zeros anyway.

References

ActiLife 6 User's Manual by the ActiGraph Software Department. 04/03/2012.

Examples

library("dplyr")
data("gtxplus1day")

gtxplus1day %>%
  collapse_epochs(60)
#> # A tibble: 1,500 x 4
#>    timestamp           axis1 axis2 axis3
#>    <dttm>              <int> <int> <int>
#>  1 2012-06-27 10:54:00  1465  1791  2572
#>  2 2012-06-27 10:55:00   207   218   270
#>  3 2012-06-27 10:56:00   169   257   270
#>  4 2012-06-27 10:57:00     0     0     0
#>  5 2012-06-27 10:58:00   157   174   248
#>  6 2012-06-27 10:59:00    23    23   279
#>  7 2012-06-27 11:00:00     0     0     0
#>  8 2012-06-27 11:01:00     0     0     0
#>  9 2012-06-27 11:02:00     0     0     0
#> 10 2012-06-27 11:03:00     0     0     0
#> # … with 1,490 more rows