Combine epochs with sleep/nonwear periods

Suppose we have used apply_tudor_locke to detect sleep periods or apply_troiano/apply_choi to detect non-wear periods. It might be useful to combine the epochs data with the periods data, so that each epoch is labeled according to which period it falls into, if any. Then we can easily slice the epochs data by sleep/non-sleep or wear/non-wear.

combine_epochs_periods(epochs, periods, start_var, end_var)

Arguments

epochs
A tibble (tbl) of activity data with (at least) a timestamp column.

periods
A summary tibble of the (sleep or non-wear) periods.

start_var
The periods column which specifies when the periods start.

end_var
The periods column which specifies when the periods end.

Value

A tibble of activity data with one additional column, period_id, which indicates the period each epoch falls into.

Examples

library("dplyr")
data("gtxplus1day")

agdb <- gtxplus1day %>%
  collapse_epochs(60) %>%
  apply_sadeh()
periods <- agdb %>%
  apply_tudor_locke(min_sleep_period = 60)

agdb_with_periods <- combine_epochs_periods(
  agdb, periods,
  in_bed_time, out_bed_time
)

# How many sleep periods were detected and what is their duration,
# in minutes?
periods %>% select(in_bed_time, out_bed_time, duration)
#>           in_bed_time        out_bed_time duration
#> 1 2012-06-27 19:34:00 2012-06-27 22:09:00      155
#> 2 2012-06-27 22:39:00 2012-06-27 23:45:00       66
#> 3 2012-06-28 00:03:00 2012-06-28 07:38:00      455
#> 4 2012-06-28 09:41:00 2012-06-28 10:53:00       72
# What is the assignment of epochs to periods?
agdb_with_periods %>% count(period_id)
#> # A tibble: 5 x 2
#>   period_id     n
#>       <int> <int>
#> 1         1   156
#> 2         2    67
#> 3         3   456
#> 4         4    73
#> 5        NA   748