Read an *.agd file, with no post-processing

Read ActiGraph sleep watch data from an SQLite database stored in an AGD file and return a list with (at least) five tables: data, sleep, filters, settings, awakenings. The tables have the schema described in the ActiLife 6 User manual and the timestamps are converted from Unix time format to human-readable POSIXct representation.

read_agd_raw(file, tz = "UTC")

Arguments

file
Full path to an agd file to read.

tz
Time zone to convert DateTime ticks to POSIX time.

Value

A list of five tables: settings, data, filters, sleep, awakenings and, if available, capsense.

Details

Some ActiGraph devices contain a capacitive sensor to detect monitor removal when worn against the skin. If that data is available, the return list includes a capsense table as well.

References

ActiLife 6 User's Manual by the ActiGraph Software Department. 04/03/2012.

covertagd: R package for converting agd files from ActiGraph into data.frames.

See also

Examples

file <- system.file("extdata", "GT3XPlus-RawData-Day01.agd",
  package = "actigraph.sleepr"
)
str(read_agd_raw(file))
#> List of 5
#>  $ data      : tibble [8,999 × 10] (S3: tbl_df/tbl/data.frame)
#>   ..$ dataTimestamp  : POSIXct[1:8999], format: "2012-06-27 10:54:00" "2012-06-27 10:54:10" ...
#>   ..$ axis1          : num [1:8999] 377 465 505 73 45 0 0 207 0 0 ...
#>   ..$ axis2          : num [1:8999] 397 816 444 91 43 0 0 218 0 0 ...
#>   ..$ axis3          : num [1:8999] 413 1225 713 106 115 ...
#>   ..$ steps          : num [1:8999] 2 4 6 1 0 0 0 1 0 0 ...
#>   ..$ lux            : num [1:8999] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ inclineOff     : num [1:8999] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ inclineStanding: num [1:8999] 10 10 10 10 0 0 0 10 0 0 ...
#>   ..$ inclineSitting : num [1:8999] 0 0 0 0 10 10 10 0 10 10 ...
#>   ..$ inclineLying   : num [1:8999] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ sleep     : tibble [0 × 10] (S3: tbl_df/tbl/data.frame)
#>   ..$ inBedTimestamp : 'POSIXct' num(0) 
#>  - attr(*, "tzone")= chr ""
#>   ..$ outBedTimestamp: 'POSIXct' num(0) 
#>  - attr(*, "tzone")= chr ""
#>   ..$ sleepID        : int(0) 
#>   ..$ timeAsleep     : int(0) 
#>   ..$ timeAwake      : int(0) 
#>   ..$ awakenings     : int(0) 
#>   ..$ wakeAfterOnset : int(0) 
#>   ..$ latency        : int(0) 
#>   ..$ efficiency     : num(0) 
#>   ..$ totalCounts    : int(0) 
#>  $ filters   : tibble [0 × 3] (S3: tbl_df/tbl/data.frame)
#>   ..$ filterStartTimestamp: 'POSIXct' num(0) 
#>  - attr(*, "tzone")= chr ""
#>   ..$ filterStopTimestamp : 'POSIXct' num(0) 
#>  - attr(*, "tzone")= chr ""
#>   ..$ filterID            : int(0) 
#>  $ settings  : tibble [33 × 3] (S3: tbl_df/tbl/data.frame)
#>   ..$ settingID   : int [1:33] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ settingName : chr [1:33] "softwarename" "softwareversion" "osversion" "machinename" ...
#>   ..$ settingValue: chr [1:33] "ActiLife" "6.13.3" "Microsoft Windows NT 10.0.14393.0" "DESKTOP-51642G4" ...
#>  $ awakenings: tibble [0 × 4] (S3: tbl_df/tbl/data.frame)
#>   ..$ timestamp  : 'POSIXct' num(0) 
#>  - attr(*, "tzone")= chr ""
#>   ..$ awakeningID: int(0) 
#>   ..$ sleepID    : int(0) 
#>   ..$ length     : int(0)