Expand time periods into a data frame of equally spaced time points

Expand time periods into a data frame of equally spaced time points

expand_periods(periods, start_var, end_var, units = "1 min")

Arguments

periods
A data frame with at least two columns, start_var and end_var, which are the first and the last epoch in a set of time periods, e.g. sleep periods or (non)wear periods.

start_var
The variable (unquoted) which indicates when the time periods start.

end_var
The variable (unquoted) which indicates when the time periods end.

units
The time unit as a characters string. The default is "1 min".

Examples

library("dplyr")
data("gtxplus1day")

gtxplus1day %>%
  collapse_epochs(60) %>%
  apply_choi(min_period_len = 45) %>%
  expand_periods(period_start, period_end, units = "30 mins")
#> # A tibble: 10 x 2
#>    period_id timestamp          
#>        <int> <dttm>             
#>  1         1 2012-06-28 00:00:00
#>  2         1 2012-06-28 00:30:00
#>  3         1 2012-06-28 01:00:00
#>  4         1 2012-06-28 01:30:00
#>  5         1 2012-06-28 02:00:00
#>  6         1 2012-06-28 02:30:00
#>  7         2 2012-06-28 04:03:00
#>  8         2 2012-06-28 04:33:00
#>  9         3 2012-06-28 06:16:00
#> 10         3 2012-06-28 06:46:00