Read activity counts from an *.agd file

Read ActiGraph sleep watch data from a database stored in an AGD file. Return a tibble.

read_agd(file, tz = "UTC")

Arguments

file
Full path to an agd file to read.

tz
Time zone to convert DateTime ticks to POSIX time.

Value

A tibble (tbl) of activity data with at least two columns: timestamp and axis1 counts. Optional columns include axis2, axis2, steps, lux and inclinometer indicators (incline off, standing, sitting and lying). The device settings are stored as attributes, which include epochlength.

References

The AGD file format is described in the ActiLife 6 Manual.

See also

Examples

file <- system.file("extdata", "GT3XPlus-RawData-Day01.agd",
  package = "actigraph.sleepr"
)
read_agd(file)
#> # A tibble: 8,999 x 10
#>    timestamp           axis1 axis2 axis3 steps   lux inclineoff inclinestanding
#>  * <dttm>              <int> <int> <int> <int> <int>      <int>           <int>
#>  1 2012-06-27 10:54:00   377   397   413     2     0          0              10
#>  2 2012-06-27 10:54:10   465   816  1225     4     0          0              10
#>  3 2012-06-27 10:54:20   505   444   713     6     0          0              10
#>  4 2012-06-27 10:54:30    73    91   106     1     0          0              10
#>  5 2012-06-27 10:54:40    45    43   115     0     0          0               0
#>  6 2012-06-27 10:54:50     0     0     0     0     0          0               0
#>  7 2012-06-27 10:55:00     0     0     0     0     0          0               0
#>  8 2012-06-27 10:55:10   207   218   270     1     0          0              10
#>  9 2012-06-27 10:55:20     0     0     0     0     0          0               0
#> 10 2012-06-27 10:55:30     0     0     0     0     0          0               0
#> # … with 8,989 more rows, and 2 more variables: inclinesitting <int>,
#> #   inclinelying <int>

library("dplyr")
library("purrr")

# Read ActiGraph sleep watch data from the AGD files in a directory
# and bind the data into one data frame indexed by `.filename`.
path <- system.file("extdata", package = "actigraph.sleepr")

list.files(path, pattern = "*.agd", full.names = TRUE) %>%
  map_dfr(read_agd, .id = ".filename")
#> # A tibble: 17,638 x 11
#>    .filename timestamp           axis1 axis2 axis3 steps   lux inclineoff
#>    <chr>     <dttm>              <int> <int> <int> <int> <int>      <int>
#>  1 1         2012-04-04 13:29:00   600  1007   883     5     0          0
#>  2 1         2012-04-04 13:29:10     0    18     8     0     0          0
#>  3 1         2012-04-04 13:29:20     0     0     0     0     0          0
#>  4 1         2012-04-04 13:29:30     0     0     0     0     0          0
#>  5 1         2012-04-04 13:29:40     0     0     0     0     0          0
#>  6 1         2012-04-04 13:29:50     0     0     0     0     0          0
#>  7 1         2012-04-04 13:30:00    55   123    40     2     0          6
#>  8 1         2012-04-04 13:30:10     0    30     0     0     0          3
#>  9 1         2012-04-04 13:30:20    22   151    18     0     0          9
#> 10 1         2012-04-04 13:30:30    11    63    16     0     0         10
#> # … with 17,628 more rows, and 3 more variables: inclinestanding <int>,
#> #   inclinesitting <int>, inclinelying <int>